Correct timeout duration to 1 hour; correct pep8 lint issues

+++ octoprint_schedule_reboot/__init__.py
@@ -27,20 +27,13 @@ class Schedule_rebootPlugin(octoprint.plugin.SimpleApiPlugin):
         log_msg = None
         if command == "schedule_reboot":
             self._logger.info("Scheduling reboot...")
-            if not self.printer_is_printing():
-                log_msg = "Printer not in use, rebooting in 60 seconds"
-                self._logger.info(log_msg)
-                self.initiate_reboot(60)
-            else:
-                log_msg = "Printer in use, will try again in 60 minutes"
-                self._logger.info(log_msg)
-                self.schedule_reboot(7200)
+            self.initial_check()
 
         elif command == "cancel":
             log_msg = "Scheduled reboot cancelled, trying again in 60 minutes"
             self._logger.info(log_msg)
             self._cancel_reboot = True
-            self.schedule_reboot(7200)
+            self.schedule_reboot(3600)
 
         elif command == "reboot_now":
             log_msg = "Reboot called immediately"
@@ -53,10 +46,21 @@ class Schedule_rebootPlugin(octoprint.plugin.SimpleApiPlugin):
     def printer_is_printing(self):
         return self._printer.is_printing() or self._printer.is_paused()
 
+    def initial_check(self):
+        if not self.printer_is_printing():
+            log_msg = "Printer not in use, rebooting in 60 seconds"
+            self._logger.info(log_msg)
+            self.initiate_reboot(60)
+        else:
+            log_msg = "Printer in use, will try again in 60 minutes"
+            self._logger.info(log_msg)
+            self.schedule_reboot(3600)
+
     def initiate_reboot(self, secs_from_now):
         """ Reboot machine in secs_from_now seconds, unless cancelled.
         """
-        if self.remaining is None:  # Only initiate a reboot if there isn't another one running
+        # Only initiate a reboot if there isn't another one running
+        if self.remaining is None:
             self._cancel_reboot = False
             self._reboot_thread = Thread(target=self._reboot_worker,
                                          args=(secs_from_now,))
@@ -65,7 +69,8 @@ class Schedule_rebootPlugin(octoprint.plugin.SimpleApiPlugin):
             payload = {'duration': secs_from_now}
             eventManager().fire(Events.SCHEDULE_REBOOT_EVENT, payload)
         else:
-            self._logger.info("Warning: Reboot scheduled while one already queued")
+            self._logger.info(
+                "Warning: Reboot scheduled while one already queued")
 
     def schedule_reboot(self, secs_from_now):
         """ Initiate a reboot in the future,
@@ -84,14 +89,21 @@ class Schedule_rebootPlugin(octoprint.plugin.SimpleApiPlugin):
                 payload = {'duration': self.remaining}
                 eventManager().fire(Events.SCHEDULE_REBOOT_EVENT, payload)
         if not self._cancel_reboot:
-            os.system('sudo reboot')
+            if not self.printer_is_printing():
+                self._logger.info("Executing reboot...")
+                os.system('sudo reboot')
+            else:
+                self._logger.info("Printer in use, will try again in 60 "
+                                  "minutes")
+                self.remaining = None
+                self.schedule_reboot(3600)
         else:
             self.remaining = None
             self._logger.info("Reboot thread aborted")
 
     def _future_reboot(self, secs_from_now):
         sleep(secs_from_now)
-        self.initiate_reboot(60)
+        self.initial_check()
 
 __plugin_name__ = "Schedule Reboot"
